Updated Makefile, also fixed conversion warnings.

+++ src/idt.c
@@ -25,7 +25,7 @@ struct idt_ptr idtp;
 void idt_set_gate(unsigned char num, unsigned long base, unsigned short sel, unsigned char flags)
 {
 	idt[num].base_lo = (base & 0xFFFF);
-	idt[num].base_hi = (base >> 16) & 0xFFFF;
+	idt[num].base_hi = (unsigned short)((base >> 16) & 0xFFFF);
 	
 	idt[num].sel = sel;
 	idt[num].always0 = 0;
